Move FileCodeWriter class into their own files
This is in preparation for creating editable files.
